Enable authentication for editing tax rates during order creation
