+++ WooCommerce/Classes/ViewRelated/Dashboard/Settings/In-Person Payments/Onboarding Errors/InPersonPaymentsPluginNotSetup.swift
@@ -55,29 +55,7 @@ struct InPersonPaymentsPluginNotSetup: View {
             .padding(.vertical, 8)
         }
         .sheet(item: $presentedSetupURL, onDismiss: onRefresh) { url in
-            setupWebView(url: url)
-        }
-    }
-
-    @ViewBuilder
-    private func setupWebView(url: URL) -> some View {
-        let stores = ServiceLocator.stores
-        let site = stores.sessionManager.defaultSite
-        if let site, stores.shouldAuthenticateAdminPage(for: site) {
-            NavigationStack {
-                AuthenticatedWebView(isPresented: .constant(true), url: url)
-                    .navigationTitle(Localization.primaryButton)
-                    .navigationBarTitleDisplayMode(.inline)
-                    .toolbar {
-                        ToolbarItem(placement: .confirmationAction) {
-                            Button(Localization.doneButton) {
-                                presentedSetupURL = nil
-                            }
-                        }
-                    }
-            }
-        } else {
-            SafariSheetView(url: url)
+            AuthenticatableWebView(url: url, title: Localization.primaryButton)
         }
     }
 
@@ -113,12 +91,6 @@ private enum Localization {
     static let refreshButton = NSLocalizedString(
         "Refresh",
         comment: "Button to refresh the state of the in-person payments setup")
-
-    static let doneButton = NSLocalizedString(
-        "inPersonPaymentsPluginNotSetup.done",
-        value: "Done",
-        comment: "Button to dismiss setup in-person payments plugin"
-    )
 }
 
 struct InPersonPaymentsPluginNotSetup_Previews: PreviewProvider {
